l10n-chile-11/requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
cryptography==2.4.2
cycler==0.9.0
defusedxml>=0.5.0<=1
dicttoxml==1.7.4
docopt==0.6.2
docutils==0.12
ebaysdk==2.1.5
erppeek
fabric
fabtools
fpdf>=1.7.2
funcsigs==0.4
future>=0.16,<0.17.0
GeoIP==1.3.2
geopy==0.95.1
gevent-websocket==0.10.1
git+https://github.com/OCA/openupgradelib.git
git+https://github.com/pysiidte/pysiidte.git@0.96
html5lib==0.999
httplib2==0.9.1
idna==2.6
IPy
Jinja2>=2.10.1
linecache2==1.0.0
lxml>=4.2.1,<5
M2Crypto==0.29.0
magento
Markdown==2.6.6
MarkupSafe==0.23
matplotlib==1.5.1
num2words==0.5.4
numpy==1.11.0
openerp-client-lib
ovh==0.4.8
passlib==1.6.5
pbr==1.8.0
pdf417gen
phonenumbers==8.9.1
Pillow>=3.3.2
psycogreen==1.0
ptyprocess==0.5
pyasn1>=0.4.1,<0.5
pycrypto>=2.6.1
pycurl==7.43.0
Pygments==2.1
pyldap
pyOpenSSL>=17.5.0
pyparsing==2.0.3
PyPDF2==1.25.1
pyserial
pysimplesoap
python-magic
python-openid
python-slugify==1.2.4
python-stdnum==1.2
python3-openid==3.0.9
pyusb
PyYAML==4.2b1
reportlab==3.3.0
roman==2.0.0
setproctitle==1.1.8
simplejson==3.8.1
six==1.11.0
zeep==3.1.0
traceback2==1.4.0
Unidecode==1.0.22
unittest2==1.1.0
urllib3>=1.24.2
vatnumber==1.2
vobject==0.9.5
Wand
Werkzeug>=0.15.3
wheel
woocommerce
XlsxWriter==0.9.6
xlrd
xlwt==1.3.0
xmltodict
signxml==2.6.0
